fluid interface for requests and better error handling
