+++ src/VaultPHP/Exceptions/KeyNameNotFoundException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace VaultPHP\Exceptions;
+
+/**
+ * Class KeyNameNotFoundException
+ * @package VaultPHP\Exceptions
+ */
+class KeyNameNotFoundException extends VaultResponseException
+{
+}
