+++ src/VaultPHP/Response/ApiErrors.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace VaultPHP\Response;
+
+class ApiErrors
+{
+    CONST ENCRYPTION_KEY_NOT_FOUND = [
+        'no existing key named .+ could be found',
+        'encryption key not found',
+    ];
+};
