+++ src/VaultPHP/SecretEngines/Engines/Transit/Request/EncryptData/EncryptDataBulkRequest.php
@@ -3,16 +3,22 @@
 namespace VaultPHP\SecretEngines\Engines\Transit\Request\EncryptData;
 
 use VaultPHP\SecretEngines\Interfaces\BulkResourceRequestInterface;
+use VaultPHP\SecretEngines\Interfaces\EncryptionTypeRequestInterface;
 use VaultPHP\SecretEngines\Interfaces\NamedRequestInterface;
 use VaultPHP\SecretEngines\Traits\ArrayExportTrait;
 use VaultPHP\SecretEngines\Traits\BulkRequestTrait;
+use VaultPHP\SecretEngines\Traits\EncryptionTypeRequestTrait;
 use VaultPHP\SecretEngines\Traits\NamedRequestTrait;
 
-final class EncryptDataBulkRequest implements BulkResourceRequestInterface, NamedRequestInterface
+final class EncryptDataBulkRequest implements
+    BulkResourceRequestInterface,
+    NamedRequestInterface,
+    EncryptionTypeRequestInterface
 {
     use ArrayExportTrait;
     use NamedRequestTrait;
     use BulkRequestTrait;
+    use EncryptionTypeRequestTrait;
 
     /**
      * EncryptDataRequest constructor.
