+++ src/VaultPHP/SecretEngines/Engines/Transit/Request/EncryptData/EncryptDataRequest.php
@@ -2,17 +2,23 @@
 
 namespace VaultPHP\SecretEngines\Engines\Transit\Request\EncryptData;
 
+use VaultPHP\SecretEngines\Interfaces\EncryptionTypeRequestInterface;
 use VaultPHP\SecretEngines\Interfaces\NamedRequestInterface;
 use VaultPHP\SecretEngines\Interfaces\ResourceRequestInterface;
+use VaultPHP\SecretEngines\Traits\EncryptionTypeRequestTrait;
 use VaultPHP\SecretEngines\Traits\NamedRequestTrait;
 
 /**
  * Class EncryptDataRequest
  * @package VaultPHP\SecretEngines\Transit\Request\EncryptData
  */
-final class EncryptDataRequest extends EncryptData implements ResourceRequestInterface, NamedRequestInterface
+final class EncryptDataRequest extends EncryptData implements
+    ResourceRequestInterface,
+    NamedRequestInterface,
+    EncryptionTypeRequestInterface
 {
     use NamedRequestTrait;
+    use EncryptionTypeRequestTrait;
 
     /**
      * EncryptDataRequest constructor.
