+++ src/VaultPHP/SecretEngines/Traits/EncryptionTypeRequestTrait.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace VaultPHP\SecretEngines\Traits;
+
+use VaultPHP\SecretEngines\Interfaces\EncryptionTypeRequestInterface;
+
+/**
+ * Trait NamedRequestTrait
+ * @package VaultPHP\SecretEngines\Traits
+ */
+trait EncryptionTypeRequestTrait
+{
+    /** @var string|null */
+    protected $type;
+
+    /**
+     * @return string|null
+     */
+    public function getType()
+    {
+        return $this->type;
+    }
+
+    /**
+     * @param $type
+     * @return $this
+     */
+    public function setType($type)
+    {
+        $this->type = $type;
+        return $this;
+    }
+}
