+++ src/VaultPHP/SecretEngines/Traits/NamedRequestTrait.php
@@ -21,10 +21,11 @@ trait NamedRequestTrait
 
     /**
      * @param string $name
-     * @return void
+     * @return $this
      */
     public function setName($name)
     {
         $this->name = $name;
+        return $this;
     }
 }
