+++ tests/VaultPHP/SecretEngines/Engines/Transit/EncryptDataTest.php
@@ -3,6 +3,7 @@
 namespace Test\VaultPHP\SecretEngines\Engines\Transit;
 
 use Test\VaultPHP\SecretEngines\SecretEngineTest;
+use VaultPHP\SecretEngines\Engines\Transit\EncryptionType;
 use VaultPHP\SecretEngines\Engines\Transit\Request\EncryptData\EncryptDataRequest;
 use VaultPHP\SecretEngines\Engines\Transit\Response\EncryptDataResponse;
 use VaultPHP\SecretEngines\Engines\Transit\Transit;
@@ -21,6 +22,7 @@ final class EncryptDataTest extends SecretEngineTest
         );
         $encryptDataRequest->setContext('fooContext');
         $encryptDataRequest->setNonce('fooNonce');
+        $encryptDataRequest->setType(EncryptionType::AES_256_GCM_96);
 
         $client = $this->createApiClient(
             'POST',
@@ -42,6 +44,7 @@ final class EncryptDataTest extends SecretEngineTest
         $this->assertEquals('foobar', $encryptDataRequest->getName());
         $this->assertEquals('fooNonce', $encryptDataRequest->getNonce());
         $this->assertEquals('fooContext', $encryptDataRequest->getContext());
+        $this->assertEquals(EncryptionType::AES_256_GCM_96, $encryptDataRequest->getType());
         $this->assertEquals(base64_encode('encryptMe'), $encryptDataRequest->getPlaintext());
     }
 }
