+++ tests/VaultPHP/VaultClientTest.php
@@ -15,6 +15,7 @@ use VaultPHP\Authentication\AuthenticationProviderInterface;
 use VaultPHP\Authentication\Provider\Token;
 use VaultPHP\Exceptions\InvalidDataException;
 use VaultPHP\Exceptions\InvalidRouteException;
+use VaultPHP\Exceptions\KeyNameNotFoundException;
 use VaultPHP\Exceptions\VaultAuthenticationException;
 use VaultPHP\Exceptions\VaultException;
 use VaultPHP\Exceptions\VaultHttpException;
@@ -30,6 +31,8 @@ use VaultPHP\VaultClient;
  */
 final class VaultClientTest extends TestCase
 {
+    use TestHelperTrait;
+
     public function testAuthProviderGetsClientInjected()
     {
         $auth = new Token('foo');
@@ -68,46 +71,6 @@ final class VaultClientTest extends TestCase
         $client->sendApiRequest('LOL', '/i/should/be/preserved', EndpointResponse::class, ['dontReplaceMe']);
     }
 
-    public function testWillThrowWhenApiHostMalformed() {
-        $this->expectException(VaultException::class);
-        $this->expectExceptionMessage('can\'t parse provided apiHost - malformed uri');
-
-        $auth = new Token('fooToken');
-        $httpClient = $this->createMock(HttpClient::class);
-
-        $client = new VaultClient($httpClient, $auth, "imInvalidHost");
-        $client->sendApiRequest('LOL', '/i/should/be/preserved', EndpointResponse::class, ['dontReplaceMe']);
-    }
-
-    public function testAuthenticateWillThrowWhenNoTokenIsReturned() {
-        $this->expectException(VaultAuthenticationException::class);
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $httpClient
-            ->expects($this->once())
-            ->method('sendRequest')
-            ->willReturn(new Response(403));
-
-        $auth = new Token("fooBar");
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        $client->sendApiRequest('GET', '/foo', EndpointResponse::class);
-    }
-
-    public function testWillThrowWhenAPIReturns403() {
-        $this->expectException(VaultAuthenticationException::class);
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $auth = $this->createMock(AuthenticationProviderInterface::class);
-        $auth
-            ->expects($this->once())
-            ->method('authenticate')
-            ->willReturn(null);
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        $client->sendApiRequest('GET', '/foo', EndpointResponse::class);
-    }
-
     public function testSendRequest() {
         $request = new Request('GET', 'foo');
         $response = new Response();
@@ -128,140 +91,9 @@ final class VaultClientTest extends TestCase
         $client->sendApiRequest('GET', '/foo',  EndpointResponse::class, [], false);
     }
 
-    public function testSendRequestWillThrow() {
-        $this->expectException(VaultHttpException::class);
-        $this->expectExceptionMessage('foobarMessage');
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $httpClient
-            ->expects($this->once())
-            ->method('sendRequest')
-            ->willThrowException(new \Exception('foobarMessage'));
-
-        $auth = $this->createMock(AuthenticationProviderInterface::class);
-        $auth->expects($this->once())
-            ->method('authenticate')
-            ->willReturn(new AuthenticationMetaData((object) [
-                'client_token' => 'foo',
-            ]));
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        $client->sendApiRequest('GET', '/foo', EndpointResponse::class);
-    }
-
-    public function testWillThrowWhenReturnClassDeclarationIsInvalid() {
-        $this->expectException(VaultException::class);
-        $this->expectExceptionMessage('Return Class declaration lacks static::fromResponse');
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $httpClient
-            ->expects($this->once())
-            ->method('sendRequest')
-            ->willReturn(new Response(200));
-
-        $auth = $this->createMock(AuthenticationProviderInterface::class);
-        $auth->expects($this->once())
-            ->method('authenticate')
-            ->willReturn(new AuthenticationMetaData((object) [
-                'client_token' => 'foo',
-            ]));
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        $client->sendApiRequest('GET', '/foo', [], MetaData::class);
-    }
-
-    public function testWillThrowWhenReturnClassDeclarationIsInvalidForBulk() {
-        $this->expectException(VaultException::class);
-        $this->expectExceptionMessage('Return Class declaration lacks static::fromBulkResponse');
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $httpClient
-            ->expects($this->once())
-            ->method('sendRequest')
-            ->willReturn(new Response(200));
-
-        $auth = $this->createMock(AuthenticationProviderInterface::class);
-        $auth->expects($this->once())
-            ->method('authenticate')
-            ->willReturn(new AuthenticationMetaData((object) [
-                'client_token' => 'foo',
-            ]));
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        $client->sendApiRequest('GET', '/foo', MetaData::class, new EncryptDataBulkRequest('foo'));
-    }
-
-    public function testWillThrowWhenResultOfReturnClassDeclarationIsInvalid() {
-        $this->expectException(VaultException::class);
-        $this->expectExceptionMessage('Result from "fromResponse/fromBulkResponse" isn\'t an instance of EndpointResponse');
-
-        $httpClient = $this->createMock(HttpClient::class);
-        $httpClient
-            ->expects($this->once())
-            ->method('sendRequest')
-            ->willReturn(new Response(200));
-
-        $auth = $this->createMock(AuthenticationProviderInterface::class);
-        $auth->expects($this->once())
-            ->method('authenticate')
-            ->willReturn(new AuthenticationMetaData((object) [
-                'client_token' => 'foo',
-            ]));
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        $client->sendApiRequest('GET', '/foo', InvalidEndpointResponseMock::class, []);
-    }
-
-    private function simulateApiResponse($responseStatus, $responseBody = '', $responseHeader = []) {
-        $response = new Response($responseStatus, $responseHeader, $responseBody);
-        $auth = new Token('fooToken');
-
-        $httpClient = $this->createMock(HttpClient::class);
-
-        $httpClient
-            ->expects($this->once())
-            ->method('sendRequest')
-            ->willReturn($response);
-
-        $client = new VaultClient($httpClient, $auth, TEST_VAULT_ENDPOINT);
-        return $client->sendApiRequest('GET', '/foo', EndpointResponse::class);
-    }
-
     public function testSuccessApiResponse() {
         $response = $this->simulateApiResponse(200, '');
         $this->assertInstanceOf(EndpointResponse::class, $response);
-    }
-
-    public function testInvalidDataResponse() {
-        $this->expectException(InvalidDataException::class);
-        $this->expectExceptionMessage('looks malformed');
-
-        $this->simulateApiResponse(400, json_encode([
-            'errors' => [
-                'looks malformed',
-            ]
-        ]));
-    }
-
-    public function testInvalidDataResponseWillConcatErrorMessages() {
-        $this->expectException(InvalidDataException::class);
-        $this->expectExceptionMessage('looks malformed, oh no');
-
-        $this->simulateApiResponse(400, json_encode([
-            'errors' => [
-                'looks malformed',
-                'oh no'
-            ]
-        ]));
-    }
-
-    public function testInvalidRouteResponse() {
-        $this->expectException(InvalidRouteException::class);
-        $this->simulateApiResponse(404, json_encode([
-            'errors' => [
-                'no handler',
-            ]
-        ]));
     }
 
     public function testEmptyResponse() {
@@ -269,22 +101,4 @@ final class VaultClientTest extends TestCase
         $this->assertInstanceOf(EndpointResponse::class, $response);
     }
 
-    public function testServerErrorResponse() {
-        $this->expectException(VaultResponseException::class);
-        $this->simulateApiResponse(500);
-    }
-
-    public function testNotHandledStatusCodeResponse() {
-        $this->expectException(VaultException::class);
-        $this->simulateApiResponse(100);
-    }
-
-    public function testResponseExceptionHasRequestResponseMeta() {
-        try {
-            $this->simulateApiResponse(555);
-        } catch (VaultResponseException $e) {
-            $this->assertInstanceOf(RequestInterface::class, $e->getRequest());
-            $this->assertInstanceOf(ResponseInterface::class, $e->getResponse());
-        }
-    }
 }
